feat: change lovelace from BigDecimal to BigInteger

+++ src/main/java/com/sotatek/cardano/common/entity/AddressTxBalance.java
@@ -1,7 +1,7 @@
 package com.sotatek.cardano.common.entity;
 
 import com.sotatek.cardano.common.validation.Word128Type;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.Objects;
 import javax.persistence.Column;
@@ -48,7 +48,7 @@ public class AddressTxBalance extends BaseEntity {
   @Column(name = "balance", nullable = false, precision = 39)
   @Word128Type
   @Digits(integer = 39, fraction = 0)
-  private BigDecimal balance;
+  private BigInteger balance;
 
   @Column(name = "time")
   private Timestamp time;
