+++ src/main/java/com/sotatek/cardano/common/entity/Epoch.java
@@ -3,7 +3,7 @@ package com.sotatek.cardano.common.entity;
 import com.sotatek.cardano.common.validation.Lovelace;
 import com.sotatek.cardano.common.validation.Word128Type;
 import com.sotatek.cardano.common.validation.Word31Type;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.Objects;
 import javax.persistence.Column;
@@ -33,12 +33,12 @@ public class Epoch extends BaseEntity {
   @Column(name = "out_sum", nullable = false, precision = 39)
   @Word128Type
   @Digits(integer = 39, fraction = 0)
-  private BigDecimal outSum;
+  private BigInteger outSum;
 
   @Column(name = "fees", nullable = false, precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal fees;
+  private BigInteger fees;
 
   @Column(name = "tx_count", nullable = false)
   @Word31Type
